refactor: separate tool execution logic from stream handling
- Extract tool execution logic into dedicated tool-execution.ts
- Simplify create-manual-tool-stream.ts to focus on stream handling
- Improve type safety and error handling
- Clean up code organization and responsibility separation

This refactoring improves code maintainability by separating
tool-related logic from stream handling, making the codebase
more modular and easier to understand.
