+++ lib/streaming/tool-execution.ts
@@ -1,0 +1,121 @@
+import {
+  CoreMessage,
+  DataStreamWriter,
+  generateId,
+  generateText,
+  JSONValue
+} from 'ai'
+import { z } from 'zod'
+import { searchSchema } from '../schema/search'
+import { search } from '../tools/search'
+import { ExtendedCoreMessage } from '../types'
+import { getToolCallModel } from '../utils/registry'
+import { parseToolCallXml } from './parse-tool-call'
+
+interface ToolExecutionResult {
+  toolCallDataAnnotation: ExtendedCoreMessage | null
+  toolCallMessages: CoreMessage[]
+}
+
+export async function executeToolCall(
+  coreMessages: CoreMessage[],
+  dataStream: DataStreamWriter,
+  model?: string
+): Promise<ToolExecutionResult> {
+  const toolCallModel = getToolCallModel(model)
+  // Convert Zod schema to string representation
+  const searchSchemaString = Object.entries(searchSchema.shape)
+    .map(([key, value]) => {
+      const description = value.description
+      const isOptional = value instanceof z.ZodOptional
+      return `- ${key}${isOptional ? ' (optional)' : ''}: ${description}`
+    })
+    .join('\n')
+
+  // Generate tool selection using XML format
+  const toolSelectionResponse = await generateText({
+    model: toolCallModel,
+    system: `You are an intelligent assistant that analyzes conversations to select the most appropriate tools and their parameters.
+            You excel at understanding context to determine when and how to use available tools, including crafting effective search queries.
+            Current date: ${new Date().toISOString().split('T')[0]}
+
+            Do not include any other text in your response.
+            Respond in XML format with the following structure:
+            <tool_call>
+              <tool>tool_name</tool>
+              <parameters>
+                <query>search query text</query>
+                <max_results>number - 5 by default</max_results>
+                <search_depth>basic or advanced</search_depth>
+                <include_domains>domain1,domain2</include_domains>
+                <exclude_domains>domain1,domain2</exclude_domains>
+              </parameters>
+            </tool_call>
+
+            Available tools: search
+
+            Search parameters:
+            ${searchSchemaString}
+
+            If you don't need a tool, respond with <tool_call><tool></tool></tool_call>`,
+    messages: coreMessages
+  })
+
+  // Parse the tool selection XML using the search schema
+  const toolCall = parseToolCallXml(toolSelectionResponse.text, searchSchema)
+
+  if (!toolCall || toolCall.tool === '') {
+    return { toolCallDataAnnotation: null, toolCallMessages: [] }
+  }
+
+  const toolCallAnnotation = {
+    type: 'tool_call',
+    data: {
+      state: 'call',
+      toolCallId: `call_${generateId()}`,
+      toolName: toolCall.tool,
+      args: JSON.stringify(toolCall.parameters)
+    }
+  }
+  dataStream.writeData(toolCallAnnotation)
+
+  // Support for search tool only for now
+  const searchResults = await search(
+    toolCall.parameters?.query ?? '',
+    toolCall.parameters?.max_results,
+    toolCall.parameters?.search_depth as 'basic' | 'advanced',
+    toolCall.parameters?.include_domains,
+    toolCall.parameters?.exclude_domains
+  )
+
+  const updatedToolCallAnnotation = {
+    ...toolCallAnnotation,
+    data: {
+      ...toolCallAnnotation.data,
+      result: JSON.stringify(searchResults),
+      state: 'result'
+    }
+  }
+  dataStream.writeMessageAnnotation(updatedToolCallAnnotation)
+
+  const toolCallDataAnnotation: ExtendedCoreMessage = {
+    role: 'data',
+    content: {
+      type: 'tool_call',
+      data: updatedToolCallAnnotation.data
+    } as JSONValue
+  }
+
+  const toolCallMessages: CoreMessage[] = [
+    {
+      role: 'assistant',
+      content: `Tool call result: ${JSON.stringify(searchResults)}`
+    },
+    {
+      role: 'user',
+      content: 'Now answer the user question.'
+    }
+  ]
+
+  return { toolCallDataAnnotation, toolCallMessages }
+}
